Added usage write when user retrieves data. Had to alter WriteMessage to take a mainTokenId which gets passed along from GetData.  Altered various Id fields to be Int64. Now recognizing select of Active field so that if a mainTokenKey is set to Active = 0 then no more data can be saved for the user.  Data can be retrieved but there will be a process that will set Bucket records to active=0 when a mainToken is inactive.

+++ LibreStore/Models/Usage.cs
@@ -3,14 +3,14 @@ namespace LibreStore.Models;
 public class Usage{
     public int Id{get;set;}
     
-    public int MainTokenId{get;set;}
+    public Int64 MainTokenId{get;set;}
     public String IpAddress{get;set;}
     
     public String Action{get;set;}
     public DateTime Created {get;set;}
     public bool Active{get;set;}
 
-    public Usage(int mainTokenId, String ipAddress, String action="")
+    public Usage(Int64 mainTokenId, String ipAddress, String action="")
     {
         MainTokenId = mainTokenId;
         IpAddress = ipAddress;
